// frontend/src/client/types.gen.ts
export type AppointmentPublic = {
    patient_name: string;
    patient_email: string;
    reason_for_visit?: (string | null);
    id: string;
    doctor_id: string;
    availability_id: string;
    appointment_time: string;
    status: AppointmentStatus;
    created_at: string;
};

export type AppointmentsPublic = {
    data: Array<AppointmentPublic>;
    count: number;
};

export type AppointmentStatus = 'pending' | 'confirmed' | 'cancelled';